+++ src/renderer/components/context-providers/VideoInfoContextsProvider.tsx
@@ -1,5 +1,5 @@
 import { createContext, ReactNode, useEffect, useReducer } from 'react';
-import { getVideoInfoInputSchema } from '../../../apiSchema';
+import { getVideoInfoInputSchema } from '../../../rpcSchema';
 import {
   VideoInfo,
   videoInfoInitialState,
